refactor(test): deduplicate scripted test fixtures via symlinks
Extract shared RegistryFixture.scala, docker.sbt, and plugins.sbt into
_fixtures/ directory. Replace 13 duplicate files across 9 test dirs with
symlinks to shared copies. RegistryFixture now registers both Order and
User schemas, so single-subject and multi-subject tests share one file.
