+++ src/sbt-test/schema-registry/_fixtures/docker.sbt
@@ -1,0 +1,8 @@
+resolvers ++= Seq("Confluent" at "https://packages.confluent.io/maven/")
+
+libraryDependencies += "org.testcontainers" % "kafka" % "1.21.3"
+
+sys.props.get("plugin.version") match {
+  case Some(v) => addSbtPlugin("org.galaxio" % "sbt-schema-registry-plugin" % v)
+  case None    => sys.error("plugin.version system property not set (provided by scriptedLaunchOpts)")
+}
