+++ src/sbt-test/schema-registry/download-failure/project/plugins.sbt
@@ -1,6 +1,1 @@
-resolvers ++= Seq("Confluent" at "https://packages.confluent.io/maven/")
-
-sys.props.get("plugin.version") match {
-  case Some(v) => addSbtPlugin("org.galaxio" % "sbt-schema-registry-plugin" % v)
-  case None    => sys.error("plugin.version system property not set (provided by scriptedLaunchOpts)")
-}
+../../_fixtures/plugins.sbt
