+++ src/sbt-test/schema-registry/download-idempotent/project/plugins.sbt
@@ -1,8 +1,1 @@
-resolvers ++= Seq("Confluent" at "https://packages.confluent.io/maven/")
-
-libraryDependencies += "org.testcontainers" % "kafka" % "1.21.3"
-
-sys.props.get("plugin.version") match {
-  case Some(v) => addSbtPlugin("org.galaxio" % "sbt-schema-registry-plugin" % v)
-  case None    => sys.error("plugin.version system property not set (provided by scriptedLaunchOpts)")
-}
+../../_fixtures/docker.sbt
