+++ src/sbt-test/schema-registry/download-success/project/RegistryFixture.scala
@@ -1,43 +1,1 @@
-import io.confluent.kafka.schemaregistry.avro.AvroSchema
-import io.confluent.kafka.schemaregistry.client.CachedSchemaRegistryClient
-import org.apache.avro.Schema
-import org.testcontainers.containers.{GenericContainer => JGenericContainer, Network}
-import org.testcontainers.containers.wait.strategy.Wait
-import org.testcontainers.kafka.ConfluentKafkaContainer
-import org.testcontainers.utility.DockerImageName
-
-/** Boots a real Confluent Schema Registry (+ Kafka) once and registers a test schema.
-  * Started lazily when `url` is first evaluated during build loading.
-  */
-object RegistryFixture {
-  val subject    = "it.e2e.Order"
-  val schemaJson =
-    """{"type":"record","name":"Order","namespace":"it.e2e","fields":[{"name":"id","type":"long"}]}"""
-
-  // Confluent image tag — keep in sync with it/src/test/.../DownloaderIntegrationSpec.scala.
-  private val confluentTag = "7.5.0"
-
-  // Containers are intentionally not stopped: the scripted JVM is short-lived and
-  // Testcontainers' Ryuk (CI) / JVM reaper (local) cleans them up on exit.
-  lazy val url: String = {
-    val network = Network.newNetwork()
-
-    val kafka = new ConfluentKafkaContainer(DockerImageName.parse(s"confluentinc/cp-kafka:$confluentTag"))
-    kafka.withListener("kafka:19092")
-    kafka.withNetwork(network)
-    kafka.start()
-
-    val sr = new JGenericContainer(DockerImageName.parse(s"confluentinc/cp-schema-registry:$confluentTag"))
-    sr.withNetwork(network)
-    sr.withExposedPorts(8081)
-    sr.withEnv("SCHEMA_REGISTRY_HOST_NAME", "schema-registry")
-    sr.withEnv("SCHEMA_REGISTRY_KAFKASTORE_BOOTSTRAP_SERVERS", "PLAINTEXT://kafka:19092")
-    sr.waitingFor(Wait.forHttp("/subjects").forStatusCode(200))
-    sr.start()
-
-    val u      = s"http://${sr.getHost}:${sr.getMappedPort(8081)}"
-    val client = new CachedSchemaRegistryClient(u, 10)
-    client.register(subject, new AvroSchema(new Schema.Parser().parse(schemaJson)))
-    u
-  }
-}
+../../_fixtures/RegistryFixture.scala
